Gate notifications, profile, security account pages behind GrowthBook flags
