+++ apps/client-fnp/app/(landing)/account/(sections)/security/SecurityClient.tsx
@@ -1,0 +1,10 @@
+"use client"
+
+export default function AccountSecurityPage() {
+    return (
+        <div className="space-y-6">
+            <h1 className="text-xl font-bold">Security</h1>
+            <p className="text-sm text-muted-foreground">Password management coming soon.</p>
+        </div>
+    )
+}
